requirements.txt
pandas
scikit-learn
mlflow
pyyaml
pytest
joblib
xgboost
flake8
